dev-requirements.txt
pylint==1.7.1
pylint-django==0.7.2